newproj/requirements.txt
# Core web and API
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3

# Networking
websockets==12.0
starlette==0.37.2

# Crypto and encoding (dev; replace with audited libs for prod)
# Use stdlib hashlib.sha3_256; remove pysha3 to avoid wheel build issues on Windows/Py3.12+.
pynacl==1.5.0
mnemonic==0.21
base58==2.1.1
argon2-cffi==23.1.0

# Database
sqlalchemy==2.0.31
alembic==1.13.2
aiosqlite==0.20.0

# Utilities
click==8.1.7
tenacity==8.3.0
colorama==0.4.6
psutil==6.0.0

# Testing
pytest==8.3.1
pytest-asyncio==0.23.7

# Windows service helpers (optional)
pywin32==306

# GUI
PySide6==6.7.2
